АТ-92 Реализация нумерации задач

+++ src/main/java/com/anst/sd/api/adapter/rest/task/dto/TaskInfoDto.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -23,6 +24,6 @@ public class TaskInfoDto {
     @NotNull
     TaskStatus status;
     String description;
-    double  orderNumber;
+    BigDecimal orderNumber;
     List<PendingNotificationDto> pendingNotifications;
 }
